+++ apps/container/src/views/Confirmation/index.ts
@@ -1,0 +1,1 @@
+export { default as ConfirmationView } from './ConfirmationView.vue';
